Fixed RRQ and WRQ
I was not transmitting them correctly. I was suppose to not know the total packets sent. Just rely on the data block being full till a zero hits meaning it is the last packet. It is fixed now for both RRQ and WRQ using the client and server (ErrorSIM should be fine with it)